requirements-dev.txt
# Дополнительные зависимости только для разработки
# Установка: pip install -r requirements-dev.txt

-r requirements.txt

# Линтер и форматирование
ruff>=0.1.0

# Покрытие кода тестами
pytest-cov>=4.0.0

# Проверка типов
mypy>=1.0.0
